src/requirements.txt
strands-agents[bidi-all]==1.28.0
boto3==1.42.1
requests==2.32.5
fastapi==0.135.1
uvicorn[standard]==0.41.0
pyaudio==0.2.14
aws-opentelemetry-distro>=0.10.1
